Enhance ScrollView to support KeyboardAware for focused inputs
