+++ packages/design-system/scroll-view/index.ts
@@ -1,0 +1,1 @@
+export { ScrollView } from "./scroll-view";
